#9 [MOO} 공통 유틸 함수 분리
